Split contrib.frontend_cache.backends.py into separate dedicated sub-modules
