+++ wagtail/contrib/frontend_cache/backends/__init__.py
@@ -1,0 +1,5 @@
+from .base import *  # noqa
+from .azure import *  # noqa
+from .http import *  # noqa
+from .cloudflare import *  # noqa
+from .cloudfront import *  # noqa
